+++ lib/honeybadger/agent.rb
@@ -128,6 +128,7 @@ module Honeybadger
       notice = Notice.new(config, opts)
 
       config.before_notify_hooks.each do |hook|
+        break if notice.halted?
         with_error_handling { hook.call(notice) }
       end
 
@@ -138,6 +139,11 @@ module Honeybadger
 
       if !opts[:force] && notice.ignore?
         debug { sprintf('ignore notice feature=notices id=%s', notice.id) }
+        return false
+      end
+
+      if notice.halted?
+        debug { 'halted notice feature=notices' }
         return false
       end