+++ CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG](http://keepachangelog.com/) for how to update this file. This project
 adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Added the ability to halt notices in callbacks using `notice.halt!`
+
 
 ## [3.3.0] - 2018-01-29
 ### Changed